requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.5
django-rest-api-logger==0.4.1
djangorestframework==3.12.2
dnspython==1.16.0
elasticsearch==7.10.1
isort==5.7.0
psycopg2-binary==2.8.6
pymongo==3.11.2
pytz==2020.5
sqlparse==0.4.1
urllib3==1.26.2
